add: database container mysql

+++ Makefile
@@ -4,6 +4,8 @@ COMPOSE_FILE = srcs/docker-compose.yml
 
 DOCKERCP += -f $(COMPOSE_FILE)
 
+SHELL=/bin/bash
+
 all :
 	$(DOCKERCP) up
 
@@ -14,14 +16,17 @@ nginx_build:
 	$(DOCKER) build --progress tty -t nginxcont srcs/nginx
 
 nginx_run:
-	$(DOCKER) run -it -p 80:80 nginxcont /bin/zsh
+	$(DOCKER) run -it -v $$PWD:/home -p 443:443 nginxcont /bin/zsh
+
+database_build:
+	$(DOCKER) build --progress tty -t databasecont srcs/database
+
+database_run:
+	$(DOCKER) run -it -v $$PWD:/home -p 443:443 databasecont /bin/zsh
 
 ips:
-	$(DOCKER) ps -aq
-	$(DOCKER) inspect -f '{{.Name}} - {{.NetworkSettings.IPAddress }}' $(docker ps -aq)
-
-php-wordpress:
-	$(DOCKER) build srcs/$@
+	$(DOCKER) ps -a
+	$(DOCKER) inspect -f '{{.Name}} - {{.NetworkSettings.IPAddress }}' $(shell docker ps -aq)
 
 list :
 	$(DOCKER) images
@@ -30,9 +35,10 @@ stop :
 	$(DOCKER) stop $(shell docker images -q)
 
 rm_all:
-	$(DOCKER) image rm -f $(shell docker images -q)
+	if [[ ! -z $$(docker ps -aq) ]];then $(DOCKER) stop $(shell docker ps -aq) ;fi
+	if [[ ! -z $$(docker images -q) ]];then $(DOCKER) image rm -f $(shell docker images -q) ;fi
 
-clr:
+clr: rm_all
 	$(DOCKER) system prune -f
 	$(DOCKER) image prune -f
 	$(DOCKER) container prune -f
